Call it Blueprinter::V2::Base
