+++ lib/blueprinter/v2/dsl.rb
@@ -4,7 +4,7 @@ require 'blueprinter/v2/association'
 require 'blueprinter/v2/field'
 
 module Blueprinter
-  class V2
+  module V2
     # Methods for defining Blueprint fields and views
     module DSL
       #
