+++ spec/v2/name_spec.rb
@@ -2,7 +2,7 @@
 
 describe "Blueprinter::V2 Names" do
   context 'const named Blueprints' do
-    class NamedBlueprint < Blueprinter::V2
+    class NamedBlueprint < Blueprinter::V2::Base
       view :extended
     end
 
@@ -28,7 +28,7 @@ describe "Blueprinter::V2 Names" do
 
   context 'manually named Blueprints' do
     let(:blueprint) do
-      Class.new(Blueprinter::V2) do
+      Class.new(Blueprinter::V2::Base) do
         self.blueprint_name = "MyBlueprint"
         view :extended
       end
@@ -47,23 +47,23 @@ describe "Blueprinter::V2 Names" do
 
   context 'anonymous Blueprints' do
     let(:blueprint) do
-      Class.new(Blueprinter::V2) do
+      Class.new(Blueprinter::V2::Base) do
         view :extended
       end
     end
 
     it 'should have no base name' do
-      expect(blueprint.blueprint_name).to eq "Blueprinter::V2"
+      expect(blueprint.blueprint_name).to eq "Blueprinter::V2::Base"
     end
 
     it 'should find a view by name' do
-      expect(blueprint[:extended].blueprint_name).to eq "Blueprinter::V2.extended"
+      expect(blueprint[:extended].blueprint_name).to eq "Blueprinter::V2::Base.extended"
     end
   end
 
   context 'deeply nested Blueprints' do
     let(:blueprint) do
-      Class.new(Blueprinter::V2) do
+      Class.new(Blueprinter::V2::Base) do
         self.blueprint_name = "MyBlueprint"
 
         view :foo do
@@ -94,7 +94,7 @@ describe "Blueprinter::V2 Names" do
   end
 
   it "should not contain periods" do
-    blueprint = Class.new(Blueprinter::V2)
+    blueprint = Class.new(Blueprinter::V2::Base)
     expect { blueprint.view :"foo.bar" }.to raise_error(
       Blueprinter::Errors::InvalidBlueprint,
       /name may not contain/
